==> R/00_doit.R <==
# Run all scripts
# ------------------------------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_analysis_and_plot.R")
source(file = "R/05_pca_and_clustering.R")
source(file = "R/06_linear_regression.R")
source(file = "R/07_ann_regression.R")
source(file = "R/08_ann_classification.R")
